Added CSSPropertyAPIMethods.json5 which defines all API methods.
Adds CSSPropertyAPIMethods.json5, which defines all methods in
CSSPropertyAPI.h. Methods are defined by name and function
definition.

This is so that when methods are added to the API, their
definitions only need to be added in this file, rather than
several places in the code.

This patch:
- Adds the CSSPropertyAPIMethods.json5 file.
- Changes make_css_property_apis.py so that it can take two files
  as input (CSSProperties.json5 and CSSPropertyAPIMethods.json5).
- Modifies the CSSPropertyAPIFiles.h.tmpl template to take this as
  input, so that it doesn't need to be modified whenever methods
  are added to the API.
- Modifies CSSPropertyDescriptor.cpp.tmpl to take this as input.

Note that this patch manually specifies all the generated output
files in build.gn, which is error-prone. Will investigate better
ways to do this in the future.

BUG=668012

patch from issue 2669243009 at patchset 320001 (http://crrev.com/2669243009#ps320001)

Review-Url: https://codereview.chromium.org/2697443002
Cr-Original-Commit-Position: refs/heads/master@{#450906}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a22b63df688ae6fb9e4101c4e81be6d22b9c555