Move metafile printing code from platform canvas to PaintCanvas
These functions operate on PaintCanvas and so move them to cc so that
they can continue to work when PaintCanvas is not an SkCanvas.

Additionally, remove skia::GetMetaData which only served to const cast a
pointer that was never const in the first place.  As the underlying
SkCanvas::getMetaData isn't const, this unconstifies the call chains.

BUG=671433
CQ_INCLUDE_TRYBOTS=master.tryserver.blink:linux_trusty_blink_rel

Review-Url: https://codereview.chromium.org/2686033005
Cr-Original-Commit-Position: refs/heads/master@{#451073}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7b64edf3a16518889d4bce514bd4bc0818e5ead2