[Android] Restart input only once on focus change
Even when finishComposingText() returns false in
InputMethodController::willChangeFocus(), we do not need to reset input
because RenderFrameImpl::focusedNodeChanged() will reset input in the
later part of focus change process.

Also, currently updateKeyboardVisibility() and updateState() in ImeAdapter
may each restart input which is another source of redundant restart input.
By merging the code path and changing some order of execution, we can
reduce additional restart input.

Lastly, we defer restartInput until the next state update. By the time we
send FocusedNodeChanged, selection is not created for the newly focused
input because of the reason mentioned in http://crbug.com/650204#c17,
so we want to defer it in order to provide correct initial sel values for
View#onCreateInputConnection().

BUG=650204

Review-Url: https://codereview.chromium.org/2681833006
Cr-Original-Commit-Position: refs/heads/master@{#451269}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62f5729e30dc50465f85419949564cb6e786dd49