Remove one navigation layout test
This CL removes http/tests/navigation/response204.html. This test checks
that a 204 response on a browser-initiated navigation will not result in
an extra item in the navigation history. This is better tested as
content_browsertest, especially since it cannot be tested using a layout
test when PlzNavigate is enabled (renderers are not made aware of a
browser-initiated navigation resulting in a 204).

BUG=689153
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/2695093006
Cr-Original-Commit-Position: refs/heads/master@{#451288}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e60da02d2e136fa7cc20abf8858b00846cbd331