Remove SafePointScopes in HeapTest.cpp
Currently SafePointScope is used for two purposes:

a) Declare that the thread is not operating on the oilpan heap
and let another thread start a GC.

b) Record the stack end and all registers before starting a GC.

Now that we have the per-thread heap, a) is not needed. b) is still needed though.
The SafePointScopes removed in this CL is a).

BUG=671856

Review-Url: https://codereview.chromium.org/2697383003
Cr-Original-Commit-Position: refs/heads/master@{#451289}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 580f0df2c038eb1e575b169ed6491e9cbe92bc11