DevTools: run user callback inside promise for promisified domains
This patch runs userCallback inside a promise. This results in errors
not being swallowed, and the promise to always resolve/reject but never
to get stuck.

As a result, the StylesSidebarPane will not die permanently
once an error happens during handling the protocol response.

BUG=681492
R=pfeldman

Review-Url: https://codereview.chromium.org/2701763002
Cr-Original-Commit-Position: refs/heads/master@{#451416}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ac15460e9aab7b68f858f911b2e4ced8e0aadaa