DevTools: teach automapping to map complex sourcemap setups.
Consider a sourcemap with inlined sources, which has one of the source
urls equals to the compiled url.

In this case, we append the ' [sm]' suffix to the source URL so that
the UISourceCode would not clash in the Workspace. (see crbug.com/562870).
This, however, kills the ability to map this file to the workspace.

This patch exploits the ability of automapping to handle query parameters.
By appending '? [sm]' instead of just ' [sm]', the automapping is able
to establish a binding based on the source size.

BUG=596422, 605548
R=dgozman

Review-Url: https://codereview.chromium.org/2698183003
Cr-Original-Commit-Position: refs/heads/master@{#451430}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9614d5a7ab4b1ea2de28ba38edb198c4e7e2a93e