Support animating <percentage> and <length-percentage> custom properties
This patch adds support for animating custom properties registered with
the <percentage> or <length-percentage> syntax.
It also fixes a bug in LengthInterpolationFunctions::createCSSValue()
when applying 0% that's unreachable without this patch.

The transition tests fail as transitions on registered custom properties
are pending implementation.

BUG=671904

Review-Url: https://codereview.chromium.org/2701533005
Cr-Original-Commit-Position: refs/heads/master@{#451558}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf1d7a3c3dc23860270f9cf267e4722b963500ab