MediaRecorder: make sure ondataavailable+onstop events are fired before onerror
This CL makes sure that "error" is the last event to be fired, after
any "dataavailable" and "stop" events have been received, which
was brought up in [1];

A typical situation triggering an error condition is if the amount
of tracks of the MediaStream being recorded changes.  Checking
for this amount of tracks is moved to the renderer/
implementation so all the errors are now sent via onError(), which
forces the recording to stop().

Added a new Layout Test for it (using <canvas>.captureStream() ISO
getUserMedia() which makes it more upstream-able).

[1] https://github.com/w3c/mediacapture-record/issues/52

BUG=693179

Review-Url: https://codereview.chromium.org/2697703003
Cr-Original-Commit-Position: refs/heads/master@{#451561}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 786184e85335e590fd0dd3b31d267f524633a90b