Allow CSSTransformInterpolationType to merge interpolated values together
This patch is part of refactoring CSS Transitions to use
CSSInterpolationTypes instead of AnimatableValues:
https://codereview.chromium.org/2680923005

When using CSSInterpolationTypes for CSS Transitions we need to be able to
use interpolated InterpolationValues as keyframe values for retargeted
composited transitions. This patch ensures CSSTransformInterpolationType
has the ability to do this.

BUG=681424

Review-Url: https://codereview.chromium.org/2693823002
Cr-Original-Commit-Position: refs/heads/master@{#451565}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8f73251b30347e8e66d536dde147da1d83b07cbb