Rename platform/wtf/ to platform/newwtf/.
This is unfortunate, but we need to give a different name to platform/wtf/
to avoid a clang-cl warning regarding MSVC-specific #include search order.

The details are explained in platform/newwtf/README.md in this change. Also,
the design doc has been amended:

https://docs.google.com/document/d/1JK26H-1-cD9-s9QLvEfY55H2kgSxRFNPLfjs049Us5w/edit?usp=sharing

BUG=691465

Review-Url: https://codereview.chromium.org/2706793002
Cr-Original-Commit-Position: refs/heads/master@{#451592}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 35820a2d012cf4d04bac9a571e8052047f336095