CSP: 'self' should work inside sandboxes.
We ought to be looking at the URL of a sandboxed resource when resolving
the CSP source expression 'self'. Currently, we're looking at the origin
of the resource, which is generally correct, but fails if the resource
has been pushed into an opaque origin.

This patch uses the fallback base URL of a document rather than its
origin to do the comparison.

BUG=692475
R=jochen@chromium.org
CC=andypaicu@chromium.org

Review-Url: https://codereview.chromium.org/2699663002
Cr-Original-Commit-Position: refs/heads/master@{#451626}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d3329c4a3d3dd0ab85869b7de4d62a8e2797520