Remove ComputedStyleBase::setBitsDefault().
ComputedStyle::setBitDefaults() resets its members to their initial
value. Since part of ComputedStyle is generated in ComputedStyleBase,
this method also calls ComputedStyleBase::setBitDefaults(), which resets
the generated members as well.

However, ComputedStyle::setBitDefaults() is only called in the
ComputedStyle constructors, so the generated members are initialised
once in the ComputedStyleBase constructor (since its a base class), and
then re-initialised a second time by a setBitDefaults() call in the
constructor body.

This patch removes ComputedStyleBase::setBitDefaults() so that
generated members are only initialised once, not twice. It also renames
ComputedStyle::setBitDefaults() to initializeBitDefaults() so it's
clearer that it's meant to be used inside the constructor.

BUG=628043

Review-Url: https://codereview.chromium.org/2699843002
Cr-Original-Commit-Position: refs/heads/master@{#451666}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f6f56fb3cadf1933dbc1b9f3419f13399287bc5d