Explicitly invalidate LayoutView when overflow is recalculated.
This bug occurs because during an animation we may change the amount of layout
overflow (via FrameView::recalcOverflowAfterStyleChange). Previously, this
would cause a change in scrollbar existance until r432965 made it so that the
FrameView would prevent scrollbar creation earlier in cases where they're known
to be unneeded. Changing scrollbar existence causes a layout so we'd implicitly
invalidate and repaint. Since that change, we (correctly) don't bother doing a
layout but because of that we no longer invalidate so the additionally exposed
background region doesn't get painted.

BUG=683814

Review-Url: https://codereview.chromium.org/2685883004
Cr-Original-Commit-Position: refs/heads/master@{#452170}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 717789e00f9ff34efe3a2ca23ebd69e802d8b6ec