Make the navigator.usb attribute accessible only in secure contexts.
Instead of performing secure context checks inside every method of the
USB object this change removes it from the Navigator object when not in
a secure context. This is equivalent to adding the [SecureContext]
extended attribute to the NavigatorUSB interface but the WebIDL code
generator doesn't support mixing it with [OriginTrialEnabled]. Instead
the secure context check is added to the hand-written Origin Trial code
in ConditionalFeaturesForModules.cpp.

BUG=691031

Review-Url: https://codereview.chromium.org/2698083008
Cr-Original-Commit-Position: refs/heads/master@{#452256}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a05ae84e268e91f39a684bc565484110c16da1b