Move tests for sendBeacon to a separate dedicated directory for clearer ownership
Not only because of the ownership clarification, but it's just good to
have a better home for them since they're closer to XHR and Fetch API.
It's not so reasonable to place them in navigation/ just because they
belongs to window.navigator.

The virtual test virtual/stable/http/tests/navigation/ is kept for a
different reason.

Moved redirection-respond.php to http/tests/resources as it's not
specific to the tests under http/tests/navigation/

BUG=682571,619662,490015,694958

Review-Url: https://codereview.chromium.org/2702113003
Cr-Original-Commit-Position: refs/heads/master@{#452408}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8a75af1cd30519bd7205f0d9129eb8db8f29c48d