Stop RenderFrameImpl from Clearing Selection During Frame Detach
When text finder stops searching without finding any match, it clears
 selection. This used to be done in RenderFrameImpl::SendFindReply,
which, however, can be called during frame detaching, causing
execCommand on a detached document.

This patch moves the call site of selection clearing to Blink, so that
we call it only when the document is still alive.

BUG=692557
TEST=content_browsertests --gtest_filter=RenderFrameImplTest.NoCrashWhenDeletingFrameDuringFind

Review-Url: https://codereview.chromium.org/2709373002
Cr-Original-Commit-Position: refs/heads/master@{#452597}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eff1b4e3dad26106e6840b0840ff3c39f7d7217f