Add UMAs for numerical approximation of table-based transfer functions.
Track the following values.

The range of the table-based transfer functions. If almost all cover the
full range from 0 to 1, then we can bake this assumption into the
numerical approxmation. This is a temporary histogram that will be
removed once we gather enough data to make an informed decision.

Whether or not the nonlinear numerical approximation converged (or if
we had to fall back to a linear approximation).

The accuracy of the final numerical approximation.

Note that none of the numerical code in this patch is used by anything
execept histograms (yet).

BUG=667966

Review-Url: https://codereview.chromium.org/2715453002
Cr-Original-Commit-Position: refs/heads/master@{#452701}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 41d5c6041911282c00f3946aa3e4759ba602434b