media: Resolve pending SetCdm() when WebMediaPlayerImpl is destructed
When WebMediaPlayerImpl is destructed, if there is a pending SetCdm()
call, we should resolve the promise so that the CDM can still be set on
the HTMLMediaElement. If we load() again, that CDM will be used as the
initial CDM for the new WebMediaPlayerImpl created.

BUG=693813
TEST=Added a new layout test.

Review-Url: https://codereview.chromium.org/2703993002
Cr-Original-Commit-Position: refs/heads/master@{#452905}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5113973bc4260b94300d81502935d5ebab57a1d0