Convert SkiaBitLocker to use PaintCanvas
SkiaBitLocker currently operates on SkCanvas, but is only used as a
helper for recording mac scrollbars.  As a part of changing out the
backend of how recordings are stored, this SkCanvas will be changed to
cc::PaintCanvas.

As this class is no longer dealing with Skia objects, rename it to
GraphicsContextCanvas and move it and its tests to platform/mac as there
is only one class that uses it.

This will allow this code to continue working when cc::PaintCanvas
is not a typedef to SkCanvas and is a separate concrete class.

BUG=671433
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2705723002
Cr-Original-Commit-Position: refs/heads/master@{#452938}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7e40b4199bd624399e821ac4dc7d0bcf2980d7b1