Worker: Pass ParentFrameTaskRunners via WorkerThread::start() instead of the ctor
This is a cleanup and should not change behavior.

ParentFrameTaskRunners is never used before WorkerThread::start() is called so
we don't have to pass it via the ctor of WorkerThread. This drastically
simplifies code and enables us to modify thread infra with a smaller number of
file changes.

BUG=694914

Review-Url: https://codereview.chromium.org/2718643002
Cr-Original-Commit-Position: refs/heads/master@{#453096}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 98eb8962d85b8c426602f89e6e31358b40e46bab