Add dummy file with very long path to test path length limits.
This CL adds a path in Chromium with a total length of 204 from the repo root, which would be the length of a file in LayoutTests/external/wpt/<subpath> where <subpath> is 160 characters.

So, if this CL doesn't break anything, then http://crrev.com/2708883004 could be committed.

This CL might break any builders with a checkout directory of about 43 characters or more.

Context: see http://crrev.com/2708883004
Review-Url: https://codereview.chromium.org/2718733002
Cr-Original-Commit-Position: refs/heads/master@{#453109}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ce0805ea777b313eb24184737d2b0dfde45e8c69