Fix specifier simplification and add test.
There was still a bug left in logic for simplifying the specifier
list for new test expectation lines: specifiers for skipped tests
were added after simplification, meaning that in the (common) case
where the test fails the same on all platforms and android is skipped,
the specifier list is simplified to [ Linux Mac Win ] and then
Android is added.

The fix is to add specifiers for skipped ports before simplifying.

BUG=655201

Review-Url: https://codereview.chromium.org/2709103009
Cr-Original-Commit-Position: refs/heads/master@{#453284}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3e6f43b480f54b4abe4b8a9e8d7e95abec31c2db