Remove inheritance on SameSizeAsComputedStyle.
Currently there is a static assert that checks the size of ComputedStyle
against a comparison struct of known size. However, the comparison
struct inherits from ComputedStyleBase, so the size of ComputedStyleBase
is hidden from view. This patch explicitly declares the size of
ComputedStyleBase in the comparison struct so that we can get a better
sense of how big ComputedStyle is.

BUG=628043

Review-Url: https://codereview.chromium.org/2716813002
Cr-Original-Commit-Position: refs/heads/master@{#453345}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 92131be1f24ca5fa037afb835c4df358269a113c